+++ fl0wer/src/main/java/org/attalos/fl0wer/rete/WorkingMemoryFinalNodeData.java
@@ -8,7 +8,7 @@ import java.util.Set;
 
 public class WorkingMemoryFinalNodeData {
     //private Set<ApplicableRule> already_fired_rules;
-    private Set<Long> already_fired_rules;
+    private Set<BigInteger> already_fired_rules;
 
     protected WorkingMemoryFinalNodeData() {
         already_fired_rules = new HashSet<>();
@@ -19,7 +19,7 @@ public class WorkingMemoryFinalNodeData {
      * @param elemIdx at which the rule got fired
      * @return true, if this rule is new (returnvalue of java.util.HashSet.add( ... ))
      */
-    protected boolean ruleFirstTimeFired(Long elemIdx) {
+    protected boolean ruleFirstTimeFired(BigInteger elemIdx) {
         return already_fired_rules.add(elemIdx);
     }
 }
